requirements.txt
Django==1.6
argparse==1.2.1
django-crispy-forms==1.4.0
requests==2.0.1
wsgiref==0.1.2
